rename curry to composeFunctions
